changes the add function to the right file, added parameters to first todo list for better readability and usage
